패키지 구조 수정에 따른 auth 파일구조 수정
