--- src/main/java/com/example/auth/dto/Login.java
@@ -1,6 +1,0 @@
-package com.example.auth.dto;
-
-public class Login {
-    String id;
-    String password;
-}
