--- src/main/java/com/example/auth/dto/User.java
@@ -1,9 +1,0 @@
-package com.example.auth.dto;
-
-public class User {
-    String id;
-    String password;
-    String name;
-    String accessToken;
-    String refreshToken;
-}
