+++ src/main/java/com/example/auth/model/User.java
@@ -1,4 +1,9 @@
 package com.example.auth.model;
 
 public class User {
+    String id;
+    String password;
+    String name;
+    String accessToken;
+    String refreshToken;
 }
